refactoring: separate authentication code from service.rb
